chore: move types to entities dir for analytics schema
